(zsh): Make path order consistent
Summary: Currently, each time a new terminal is open we add various things to
the PATH.  Apple's `path_helper` in `zprofile` messes up the order of entries
in PATH.  This commit only sets the PATH if we detect it hasn't been set yet.

Note: This makes it harder to apply changes to `paths.sh` in the future,
because we cannot simply open a new Tmux window.