fix(randomcowcommand): workaround bug in macOS' `man` command
Summary:

The `man` command on macOS has a bug where it doesn't correctly import `.so`
files when you give it the path to a file. This is because it `cd`s to the
wrong directory for the relative imports to work.

This commit works around the bug by doing what `man FILEPATH` does, but in the
right directory.